requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
sqlalchemy==2.0.38
pydantic==2.10.6
python-jose==3.4.0
passlib==1.7.4
redis==5.2.1
celery==5.4.0
requests==2.32.3
openai==1.64.0